Move preparams storage code to preparams.go
Merged the files as the code is related and the files are not that long.
